add spider disappear test
