+++ playwright/tests/detail_page/maps/test_detail_map_interactions.py
@@ -33,7 +33,9 @@ def test_drawing_shape_adds_download_filter(
     )
     detail_page.mouse.move(x, y)
     detail_page.detail_map.click_map()
-    expect(detail_page.bbox_condition_box.first).to_have_css("visibility", "visible", timeout=5000)
+    expect(detail_page.bbox_condition_box.first).to_have_css(
+        'visibility', 'visible', timeout=5000
+    )
 
     # Remove the drawn shape
     detail_page.detail_map.hover_map()
